modules/zantara-deploy/requirements.txt
# Zantara Orchestrator Dependencies

# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Security
PyJWT>=2.8.0
python-multipart>=0.0.6

# Environment management
python-dotenv>=1.0.0

# Google Cloud (optional, for production)
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.13.0
google-cloud-secret-manager>=2.17.0
google-cloud-logging>=3.8.0

# HTTP requests (for testing)
requests>=2.31.0

# Development
pytest>=7.4.0
httpx>=0.25.0